Rename config to configs
