+++ qiskit_experiments/framework/base_experiment.py
@@ -27,7 +27,7 @@ from qiskit.qobj.utils import MeasLevel
 from qiskit.providers.options import Options
 from qiskit_experiments.framework.store_init_args import StoreInitArgs
 from qiskit_experiments.framework.experiment_data import ExperimentData
-from qiskit_experiments.framework.config import ExperimentConfig
+from qiskit_experiments.framework.configs import ExperimentConfig
 
 
 class BaseExperiment(ABC, StoreInitArgs):
